Intial commit minus graphics files
The rubric of radiative transfer covers a fascinating range physical processes includ- ing astrophysics, planetary sciences, and climate modeling. As such, purview of this research is restricted to a pressing experimental problem: inertial confinement physics, in particular the thermonuclear burn of a deuterium-tritium plasma at the National Ignition Facility at Lawrence Livermore National Laboratory. Even so, the problem demands a tightened focus on magnetohydrodynamic instabilities in the imploding plasma. Here we focus on the Rayleigh-Taylor instability caused by the impulsive driving of the outer core of capsule. The goal is to understand how radiation transported from the outer edge of the capsule signals the thermodynamic behavior of the mixing materials.